Only fix FPM configs if neeeded

+++ docker/domserver/scripts/start.sh
@@ -82,8 +82,14 @@ cp etc/nginx-conf /etc/nginx/sites-enabled/default
 # Replace nginx php socket location
 sed -i 's/server unix:.*/server unix:\/var\/run\/php-fpm-domjudge.sock;/' /etc/nginx/sites-enabled/default
 # Remove default FPM pool config and link in DOMJudge version
-rm /etc/php/7.0/fpm/pool.d/www.conf
-ln -s /opt/domjudge/domserver/etc/domjudge-fpm.conf /etc/php/7.0/fpm/pool.d/domjudge.conf
+if [[ -f /etc/php/7.0/fpm/pool.d/www.conf ]]
+then
+	rm /etc/php/7.0/fpm/pool.d/www.conf
+fi
+if [[ ! -f /etc/php/7.0/fpm/pool.d/domjudge.conf ]]
+then
+	ln -s /opt/domjudge/domserver/etc/domjudge-fpm.conf /etc/php/7.0/fpm/pool.d/domjudge.conf
+fi
 # Change pm.max_children
 sed -i "s/^pm\.max_children = .*$/pm.max_children = ${FPM_MAX_CHILDREN}/" /etc/php/7.0/fpm/pool.d/domjudge.conf
